Refactor PersonalAccessTokenExpiry and MessageExpiry utils
This commit moves the PersonalAccessTokenExpiry and MessageExpiry
utilities from the cli/utils module to the utils module. It updates
all references across the cli and sdk modules to reflect this change.
The commit also removes the now redundant cli/utils module files.

This fixes #789
