+++ cli/src/args/personal_access_token.rs
@@ -1,6 +1,6 @@
 use crate::args::common::ListMode;
 use clap::{Args, Subcommand};
-use iggy::cli::utils::personal_access_token_expiry::PersonalAccessTokenExpiry;
+use iggy::utils::personal_access_token_expiry::PersonalAccessTokenExpiry;
 use std::convert::From;
 
 #[derive(Debug, Clone, Subcommand)]
