+++ sdk/src/binary/personal_access_tokens.rs
@@ -1,7 +1,6 @@
 use crate::binary::binary_client::{BinaryClient, BinaryClientV2};
 use crate::binary::{fail_if_not_authenticated, mapper, BinaryTransport, ClientState};
 use crate::bytes_serializable::BytesSerializable;
-use crate::cli::utils::personal_access_token_expiry::PersonalAccessTokenExpiry;
 use crate::client::PersonalAccessTokenClient;
 use crate::client_v2::PersonalAccessTokenClientV2;
 use crate::command::*;
@@ -12,6 +11,7 @@ use crate::personal_access_tokens::create_personal_access_token::CreatePersonalA
 use crate::personal_access_tokens::delete_personal_access_token::DeletePersonalAccessToken;
 use crate::personal_access_tokens::get_personal_access_tokens::GetPersonalAccessTokens;
 use crate::personal_access_tokens::login_with_personal_access_token::LoginWithPersonalAccessToken;
+use crate::utils::personal_access_token_expiry::PersonalAccessTokenExpiry;
 
 #[async_trait::async_trait]
 impl<B: BinaryClient> PersonalAccessTokenClient for B {
