+++ sdk/src/cli/utils/mod.rs
@@ -1,3 +1,1 @@
 pub mod login_session_expiry;
-pub mod message_expiry;
-pub mod personal_access_token_expiry;
