+++ sdk/src/http/personal_access_tokens.rs
@@ -1,4 +1,3 @@
-use crate::cli::utils::personal_access_token_expiry::PersonalAccessTokenExpiry;
 use crate::client::PersonalAccessTokenClient;
 use crate::client_v2::PersonalAccessTokenClientV2;
 use crate::error::IggyError;
@@ -10,6 +9,7 @@ use crate::personal_access_tokens::create_personal_access_token::CreatePersonalA
 use crate::personal_access_tokens::delete_personal_access_token::DeletePersonalAccessToken;
 use crate::personal_access_tokens::get_personal_access_tokens::GetPersonalAccessTokens;
 use crate::personal_access_tokens::login_with_personal_access_token::LoginWithPersonalAccessToken;
+use crate::utils::personal_access_token_expiry::PersonalAccessTokenExpiry;
 use async_trait::async_trait;
 
 const PATH: &str = "/personal-access-tokens";
