+++ sdk/src/http/topics.rs
@@ -1,4 +1,3 @@
-use crate::cli::utils::message_expiry::MessageExpiry;
 use crate::client::TopicClient;
 use crate::client_v2::TopicClientV2;
 use crate::error::IggyError;
@@ -13,6 +12,7 @@ use crate::topics::get_topics::GetTopics;
 use crate::topics::purge_topic::PurgeTopic;
 use crate::topics::update_topic::UpdateTopic;
 use crate::utils::byte_size::IggyByteSize;
+use crate::utils::message_expiry::MessageExpiry;
 use async_trait::async_trait;
 
 #[async_trait]
