+++ sdk/src/utils/personal_access_token_expiry.rs
@@ -1,0 +1,3 @@
+use crate::utils::message_expiry::MessageExpiry;
+
+pub type PersonalAccessTokenExpiry = MessageExpiry;
